+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `safe_pool`:
     - All pool methods of all pool instances now supports keyword-only arguments `args` and `kwds`, which allows passing
       arguments to the called function in the same way as in `apply`.
+    - Add `broadcast` method for stateful pools.
 - `take` and `drop` iterator methods.
 
 ### Changed